More work on the new OpenCL options
Now the new options show up in the "Detailed Calculation Settings" dialog and
are saved and restored from the per-user configuration.

The code that manipulates the "Detailed Calculation Settings" dialog is quite
ugly with all its manual hiding and showing of widgets depending on which
detail it is that is being edited. This also means that the dialog cannot be
designed using Glade. But no time now to re-work this.

Change-Id: I03a3a51d902084e73aab5a787b588d22ea7578f2