Synchronize bullets lists
The default bullets list is defined in two more places:
svx/source/dialog/svxbmpnumvalueset.cxx
cui/source/tabpages/numpages.cxx
in both cases there is no separate RTL list. (And since we're
using SvxNumValueSet also for the popup, we must sync it,
so the shown bullet would be always the applied one.)

There is lots of code similarity/duplication in this area
(primarily introduced by the sidebar), and I hope to clean this
at some point.

Change-Id: Ifcc216101ed2f105a40fb483aa1b96a9e93d00e0