windows opengl: Proof-of-concept text rendering.
We don't have a method that would paint a texture with transparency yet.  We
also need to limit the size of the DIBSection exactly to what we are going to
paint, no point in creating a huge bitmap that is mostly empty (but the part
where is the text being drawn).

Change-Id: Ice0bf325743d08e19e636be73cef6aff3cde5704