DOCX export: put either hints or the para attr set into the para marker props
Commit 99bb0097feb1fb099988e51dc0bc57f0d1017754 (abi#9930
DocxAttributeOutput::AddToAttrList: avoid duplicated attributes,
2014-11-02) added a quick fix to avoid writing duplicated attributes,
but the validation problem of duplicated elements remained.

Make sure that when writing paragraph marker properties, then we write
either the hints at the end of the paragraph or the character attributes
range of the paragraph properties, but not both.

The only exception is the character style that's always set as
a hint, as explained in commit af9c43455aeaef3c7eb38e530a377b3c633e4c4c
(DOCX export: fix handling of paragraph mark on empty paragraphs,
2014-10-17).

Change-Id: I494e5bb9871aa535532fef32bd344d8093e9f762