Fix rtf outline importing issue.
When importing RTF files, the outline is treated
as normal numbering list. Open the tools>chapter, outline
doesn't correspond to heading styles correctly.

This patch is to handle RTF keyword \s in a \listlevel. 
The patch use style name as its id so that is consistent 
with the one already used by StyleSheetEntry.

Change-Id: I5ab6602b5ce64ca65ec08e0adb34d60ae7293650
Reviewed-on: https://gerrit.libreoffice.org/12960
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Miklos Vajna <vmiklos@collabora.co.uk>