DOCX import: first page header should always set default headers as well
writerfilter imports a section with first and default headers / footers
as two page styles: a first page one and a follow one. In case we
nominally have a header on the first page only, we still need to enable
it in the follow page style as well, otherwise on export when we
recognize the first + follow chain as a single section, we'll end up
with a header that looses its header margin.

Change-Id: I4923daffefb243d2de4805a9931150ce4d9a6c81