fdo#86573: Impress does not show pictures anymore
Regression from:
07c975a2981f3460a63a3d063d8e6c8d67428db3

Sometimes SdrGrafObj import the image file just as a
preview (with low resolution) and late when the real
picture is needed it just throws this preview and tries
to load the image again.
This preview works only with package URL so in this
case we should not forget that URL.

Plus two other related potential image loss situation:
- SetUserData() is called with empty image URL
- Image preview is thrown when we actually have no
package URL from we can import the image again.

Change-Id: Ib21f400eb6d28133e8598e6b201209e5dd1a976f