We now only need one bit of information here
whether there *is* an id, or whether there is not an id.  An id of 0 seems to
be used as a flag that an interface is some sort of superclass which can be
subclassed by something else and have its toolbars etc reused.

Convert this to a bool so we don't need an a resource id for each one and just
drop the third arg for the normal "final" case and use a different define for
the inheritable case

Change-Id: I98380f03d73d57bf8cba02d339097e384518abaa