Unittest for autocorrect in writer
At first we test normal replacement of wrong words.
Second part of the test is about style changes during autocorrection.
And at the end we replace a word with a table.

The rules for the autocorrection are in
test/user-template/user/autocorr/acor_fr.dat
I needed to make SwAutoCorrect public for my test.

Conflicts:
	sw/qa/extras/uiwriter/uiwriter.cxx

Change-Id: Ie753c2187c1f4eddaefe0cc428a605107112c24f
Reviewed-on: https://gerrit.libreoffice.org/13131
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>