fdo#86117 a11y: Exception when closing popup from ToolBox w/o separator
For some reason PopupWindowControllerImpl::WindowEventListener is getting
the VCLEVENT_WINDOW_HIDE event twice, and while processing the second one
the ToolBox no longer has an active down item, which results in a
IndexOutOfBoundsException in VCLXAccessibleToolBox::getAccessibleChild,
because ToolBox::GetItemPos(0) correctly returns TOOLBOX_ITEM_NOTFOUND.
But when a ToolBox has at least one separator, ToolBox::GetItemPos(0)
"detects" its position as the current down item, and no exception is
thrown.

Probably it just hides the bug, because it seems to me that getting
the hide event twice is the real bug here.

Change-Id: If018350dd91cd959c0c8f7d6859474f95fb8cd1e
Reviewed-on: https://gerrit.libreoffice.org/13173
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>
Tested-by: Maxim Monastirsky <momonasmon@gmail.com>