fdo#72031 RTF import: bogus call to getBestTextEncodingFromLocale()
There were two problems here. First, commit
a3aeb1463e2f8c711dd642cf7cbaa39e6a78b729 (rtftok: stop sending
sprm:CRgFtc{0,1,2} tokens, 2014-03-05) broke the use-case when the font
encoding is 0, but it's present. Before that commit, we parsed the font
encoding instantly; after that commit we parse it once we have a font
name. If we do that, then we have to have an idea if we have a font
encoding. Given that 0 is a valid encoding, use -1 for the "have no
encoding" case instead.

Second, commit ec7c6ca4d54cc60f8228a6cd6423592ddfae2f6e (fdo#85889
handle pc, pca and mac rtf keywords in writerfilter, 2014-11-24) abused
m_nCurrentEncoding, which is meant to be used within the font table
only. The problem with this is that this way only the first font will
get the encoding, while the spec says it should be used in every context
where there is no other explicit encoding. Fix this by setting the
default encoding for those 3 control words instead -- and consider the
default encoding in getEncoding().

Change-Id: Ia1d71f8ce70f2a53a3770b4840e21362d082e71f