fdo#87005 SwTxtFormatter::NewPortion: fix lost SwFlyPortion
In case there are two flys anchored to the paragraph and it's empty,
then the situation is that SwTxtFormatter::BuildPortions() calls
NewPortion() two times, second time the SwTxtFormatInfo has no fly set
initially, only after NewPortion() calls CalcFlyWidth(). When that
happens, we used to return pPor, even in case it was 0, and we the
SwTxtFormatInfo had a fly portion.

Fix the problem by checking if SwTxtFormatInfo has a fly portion after
CalcFlyWidth(), and in case otherwise we would return 0, return the fly
portion of SwTxtFormatInfo instead. As a result, the paragraph delimiter
in the bugdoc will be positioned at the correct position, as the
SwTxtFrm will properly have two layout portions, just like non-empty
SwTxtFrms.

Change-Id: I51e5ba61e79b4353c7b11c6d76b8c370ac3d4d37