Accessibility: Expose keyboard short-cuts of menu entries
On Windows the keyboard short-cuts of menu items like
"Open Ctrl+O" is not exposed to accessibility tools on
Windows. On Mac and Linux it is.

If I understand Microsoft's notes correctly, it is expected
that the short-cut gets appended to the accessibleName.
This is what Acrobat Reader, Internet Explorer, Firefox
is doing. And with this patch LibreOffice as well.

For info from Microsoft on the matter see:
http://msdn.microsoft.com/en-us/library/windows/desktop/dd318482%28v=vs.85%29.aspx

Change-Id: I4ed7c4c45998063bb50744b44c792351070489d7
Reviewed-on: https://gerrit.libreoffice.org/13516
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>