bnc#830738 Now all axis labels of a chart are displaied.
Problem:
Some axis labels of a chart are missing, when they should not.

Analysis and solution:
The current implementation is affected by the following issues:

1) When the method switches to the 45-degrees layout the
`nRhythm` parameter is not reset to 1.
2) The bounding boxes computed by the `doesOverlap` test routine
for the 45-degrees layout are wrong.

Because of the first issue only one label every `nRhythm` is showed
even if in the 45-degrees layout no overlap occurs.

The second issue is located inside the `lcl_getRotatedPolygon`
routine which is used by the `doesOverlap` routine for building a
polygon representing the bb for a text label. The polygon is created
in the following way: a new rectangle is created: the top-left vertex is
placed at axes origin and it is initialized with the same width and
height of the bb of the passed text label. Later several
transformations are performed on the new rectangle:
1) the rectangle is rotated by a 45-degrees angle respect with its
center; 2) the rectangle is translated by a vector equals to the
top-left vertex of the bb of the passed text label.
There are 2 errors in this sequence of transformations:
1) The `B2DHomMatrix` class used for representing the
transformation, performs rotations in the positive direction (from
the X axis to the Y axis). However since the coordinate system used
by the chart has the Y-axis pointing downward, a rotation in the
positive direction means a clockwise rotation. On the contrary text
labels are rotated counterclockwise. This can be easily fixed by
using the opposite angle. 2) Rotating the rectangle respect with its
center is wrong since the fixed point of the rotation must be the
top-left corner of the rectangle: in this way when the final
translation is performed the top-left vertex of the transformed
rectangle will be coincident with the top-left vertex of the text label
bb, and so the transformed rectangle will be coincident with the text
label bb.

Change-Id: Ia8072c3d5bd6c1a11f66c33d45d56e55634edf4c