Regression in rendering of flowChartPunchedTape custom shape
Regression from:
a2c65d5177dc998bab0fc6125948f62a4b71db7d

It's not a good idea to create the arc first in small
resolution and scale it later, because the result won't
be smooth (e.g. in case of flowChartPunchedTape).

Both cloud and flowChartPunchedTape rendered well after
that change. It seems the problem in case of cloud
was with the arc tangent values which should be calculated
based on the ARCANGLETO's original values before scaling.

Change-Id: I0315afd0452108b6539b4b8ba271d1615810d985