No need for USE_DOUBLE_MMAP on OS X
I had introduced it with 4106e8bd6999bfb74f339f150f0a22b89cc11333 in
2012, when the C++/UNO bridge for 64-bit OS X was created mostly as a
copy of the Linux one.

Stephan says that the only need for USE_DOUBLE_MMAP should be on
SELinux anyway, so most likely also its use for the various BSDs and
Android are copypasta or cargo cult.

Change-Id: I1c16e830e5e8269b78b14837a9127a98612a6e54