do not set Y line width to 0
I'm somewhat confused by why there needs to be a separate line width
for X and Y, but apparently there is, and the latter shouldn't be just plain 0
(otherwise a number of drawPolyLine() implementations either skip using
a simpler path for the usual case of them being equal, or even plain to refuse
work at all and cause a fall back).
And I hope this doesn't lead to finding out that some of those implementation
are actually buggy.

Change-Id: I2dbbd1539c4a96d41935cce9ae6565872e2a459b