do not drop entire external reference, fdo#85617 related
If there are no matching tab names for a FileId, preserve at least the known
reference parts. In case of 2D references the sheet name is in the token if
for example read from .xlsx, only for 3D references the second sheet name
would be needed. The underlying makeExternalRefStr() and its subroutines
handle the missing tabname elements gracefully.

Still this situation is worth an assertion.