gbuild: try to fix windows non-debug builds
The ENABLE_CRASHREP is no longer AC_SUBSTed by configure, so it is
always non-empty in config_host.mk, resulting in -Zi without -FS,
which breaks MSVC 2013 builds.

Leave a placeholder in configure for the new crash reporter vaporware,
and force the CFLAGS to be the same as in the --eanble-symbols case.

(regression from 7b322d8d69dcff943c3a4313980958ecee86424c)

Change-Id: I476fb2e4f50de5304632628b25e3997ea6c31c27