Android: Abstract file management from LibreOfficeUIActivity.
This is the first step towards implementing support for different
cloud providers.

The use of File objects is replaced with the interface IFile, which
has similar operations. A LocalFile implementation is provided to
deal with local contents in the same way it was being done before.

Some system-wide storage operations are abstracted in the interface
IDocumentProvider. A LocalDocumentsProvider implementation is
provided.

Known issues in this patch:

* Dummy filesystem code has been removed. This should be provided as
  an implementation of IDocumentProvider in case we need it, but that
  will probably not happen.

* Sorting is not implemented in LocalDocumentsProvider yet. It seemed
  not to be working anyway.

* Code to get directory thumbnails has been commented out. This code
  will probably belong to the IDocumentProvider/IFile implementations.
  We might want to have a different approach instead of reading
  hidden thumbnail files, something similar to what the native
  Start Center does.

Change-Id: Ib0882c69a0b24e28d2346bbe0154db01d1102b06